third_party/blink/web_tests/http/tests/devtools/profiler/heap-snapshot-summary-show-ranges-expected.txt

Tests showing several node ranges in the Summary view of detailed heap snapshot.

Profiler was enabled.
Detailed heap profiles were enabled.

Running: testShowAll
Row found: true

Step 3
Retrieved ranges: [{"from":0,"to":10}]
[0] A 773492  492  
[1] A 213482  482  
[2] A 613482  482  
[3] A 853472  472  
[4] A 1813462  462  
[5] A 413452  452  
[6] A 1413452  452  
[7] A 1853442  442  
[8] A 573432  432  
[9] A 533412  412  
[10] Show 10 beforeShow all 40Show 10 after

Step 4
Retrieved ranges: [{"from":0,"to":10},{"from":30,"to":40}]
[0] A 773492  492  
[1] A 213482  482  
[2] A 613482  482  
[3] A 853472  472  
[4] A 1813462  462  
[5] A 413452  452  
[6] A 1413452  452  
[7] A 1853442  442  
[8] A 573432  432  
[9] A 533412  412  
[10] Show 10 beforeShow all 20Show 10 after
[11] A 1333241  241  
[12] A 173231  231  
[13] A 1453231  231  
[14] A 1773211  211  
[15] A 1973211  211  
[16] A 1893181  181  
[17] A 253161  161  
[18] A 1093151  151  
[19] A 2013151  151  
[20] A 1933131  131  
[21] Show 10 beforeShow all 10Show 10 after

Step 5
Retrieved ranges: [{"from":0,"to":10},{"from":20,"to":25},{"from":30,"to":40}]
[0] A 773492  492  
[1] A 213482  482  
[2] A 613482  482  
[3] A 853472  472  
[4] A 1813462  462  
[5] A 413452  452  
[6] A 1413452  452  
[7] A 1853442  442  
[8] A 573432  432  
[9] A 533412  412  
[10] Show 10 beforeShow all 10Show 10 after
[11] A 1293342  342  
[12] A 813312  312  
[13] A 1573302  302  
[14] A 1373291  291  
[15] A 1693291  291  
[16] Show 10 beforeShow all 5Show 10 after
[17] A 1333241  241  
[18] A 173231  231  
[19] A 1453231  231  
[20] A 1773211  211  
[21] A 1973211  211  
[22] A 1893181  181  
[23] A 253161  161  
[24] A 1093151  151  
[25] A 2013151  151  
[26] A 1933131  131  
[27] Show 10 beforeShow all 10Show 10 after

Step 6
Retrieved ranges: [{"from":0,"to":10},{"from":20,"to":25},{"from":28,"to":40}]
[0] A 773492  492  
[1] A 213482  482  
[2] A 613482  482  
[3] A 853472  472  
[4] A 1813462  462  
[5] A 413452  452  
[6] A 1413452  452  
[7] A 1853442  442  
[8] A 573432  432  
[9] A 533412  412  
[10] Show 10 beforeShow all 10Show 10 after
[11] A 1293342  342  
[12] A 813312  312  
[13] A 1573302  302  
[14] A 1373291  291  
[15] A 1693291  291  
[16] Show 10 beforeShow all 3Show 10 after
[17] A 453251  251  
[18] A 333241  241  
[19] A 1333241  241  
[20] A 173231  231  
[21] A 1453231  231  
[22] A 1773211  211  
[23] A 1973211  211  
[24] A 1893181  181  
[25] A 253161  161  
[26] A 1093151  151  
[27] A 2013151  151  
[28] A 1933131  131  
[29] Show 10 beforeShow all 10Show 10 after

Step 7
Retrieved ranges: [{"from":0,"to":10},{"from":18,"to":26},{"from":28,"to":40}]
[0] A 773492  492  
[1] A 213482  482  
[2] A 613482  482  
[3] A 853472  472  
[4] A 1813462  462  
[5] A 413452  452  
[6] A 1413452  452  
[7] A 1853442  442  
[8] A 573432  432  
[9] A 533412  412  
[10] Show 10 beforeShow all 8Show 10 after
[11] A 1253352  352  
[12] A 1173342  342  
[13] A 1293342  342  
[14] A 813312  312  
[15] A 1573302  302  
[16] A 1373291  291  
[17] A 1693291  291  
[18] A 2053281  281  
[19] Show 10 beforeShow all 2Show 10 after
[20] A 453251  251  
[21] A 333241  241  
[22] A 1333241  241  
[23] A 173231  231  
[24] A 1453231  231  
[25] A 1773211  211  
[26] A 1973211  211  
[27] A 1893181  181  
[28] A 253161  161  
[29] A 1093151  151  
[30] A 2013151  151  
[31] A 1933131  131  
[32] Show 10 beforeShow all 10Show 10 after

Step 8
Retrieved ranges: [{"from":0,"to":10},{"from":15,"to":45}]
[0] A 773492  492  
[1] A 213482  482  
[2] A 613482  482  
[3] A 853472  472  
[4] A 1813462  462  
[5] A 413452  452  
[6] A 1413452  452  
[7] A 1853442  442  
[8] A 573432  432  
[9] A 533412  412  
[10] Show 10 beforeShow all 5Show 10 after
[11] A 653382  382  
[12] A 1133382  382  
[13] A 1533362  362  
[14] A 1253352  352  
[15] A 1173342  342  
[16] A 1293342  342  
[17] A 813312  312  
[18] A 1573302  302  
[19] A 1373291  291  
[20] A 1693291  291  
[21] A 2053281  281  
[22] A 373271  271  
[23] A 493261  261  
[24] A 453251  251  
[25] A 333241  241  
[26] A 1333241  241  
[27] A 173231  231  
[28] A 1453231  231  
[29] A 1773211  211  
[30] A 1973211  211  
[31] A 1893181  181  
[32] A 253161  161  
[33] A 1093151  151  
[34] A 2013151  151  
[35] A 1933131  131  
[36] A 1493121  121  
[37] A 693111  111  
[38] A 29390  90  
[39] A 121380  80  
[40] A 165380  80  
[41] Show 10 beforeShow all 5Show 10 after

Step 9
Retrieved ranges: [{"from":0,"to":45}]
[0] A 773492  492  
[1] A 213482  482  
[2] A 613482  482  
[3] A 853472  472  
[4] A 1813462  462  
[5] A 413452  452  
[6] A 1413452  452  
[7] A 1853442  442  
[8] A 573432  432  
[9] A 533412  412  
[10] A 893402  402  
[11] A 973402  402  
[12] A 733392  392  
[13] A 1013392  392  
[14] A 133382  382  
[15] A 653382  382  
[16] A 1133382  382  
[17] A 1533362  362  
[18] A 1253352  352  
[19] A 1173342  342  
[20] A 1293342  342  
[21] A 813312  312  
[22] A 1573302  302  
[23] A 1373291  291  
[24] A 1693291  291  
[25] A 2053281  281  
[26] A 373271  271  
[27] A 493261  261  
[28] A 453251  251  
[29] A 333241  241  
[30] A 1333241  241  
[31] A 173231  231  
[32] A 1453231  231  
[33] A 1773211  211  
[34] A 1973211  211  
[35] A 1893181  181  
[36] A 253161  161  
[37] A 1093151  151  
[38] A 2013151  151  
[39] A 1933131  131  
[40] A 1493121  121  
[41] A 693111  111  
[42] A 29390  90  
[43] A 121380  80  
[44] A 165380  80  
[45] Show 10 beforeShow all 5Show 10 after

Profiler was disabled.
